refactor(direct-control): extract operation router source
Move the embedded generic operation validation/send router source into packages/civ7-direct-control/src/play/operations/router.ts and import it from the package facade. Keep wrapper-level postcondition classifiers and specialized production, diplomacy, narrative, and unit-target closeout sources in index.ts for later owner slices.

Record the partial operation atom status in the support modularization OpenSpec workstream. This is local package source relocation proof only, not runtime or live-game proof.

Validation: bun run --cwd packages/civ7-direct-control test -- unit-operation.test.ts production-choice.test.ts diplomacy-response.test.ts narrative-choice.test.ts unit-target-action.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
