+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -126,6 +126,9 @@ runtime/direct-control claims.
   - [x] 4.5.3 Extract ready-city embedded source owner while keeping the public
         wrapper in the facade.
 - [ ] 4.6 Extract operation validation/send/postcondition atoms.
+  - [x] 4.6.1 Extract operation router embedded validation/send source owner
+        while keeping wrapper-level postconditions and specialized closeouts in
+        the facade.
 - [x] 4.7 Extract settlement/tactical/progression read atoms.
   - [x] 4.7.1 Extract settlement recommendation embedded source owner while
         keeping the public wrapper in the facade.
